chore: refactor analysis module structure
