+++ src/analysis/mod.rs
@@ -1,16 +1,12 @@
-pub mod compare;
-pub mod executor;
-use compare::{Comparison, ComparisonResult};
-use executor::{ExecutionContext, ExecutionError};
-use std::path::Path;
-use std::time::Duration;
-use thiserror::Error;
+mod compare;
+mod context;
+mod executor;
+mod result;
 
-pub struct AnalysisContext {
-    pub testname: String,
-    pub gccrs: ExecutionContext,
-    pub rustc: ExecutionContext,
-}
+pub use executor::ExecutionError;
+pub use context::AnalysisContext;
+pub use thiserror::Error;
+pub use result::AnalysisResult;
 
 #[derive(Debug, Error, PartialEq)]
 pub enum AnalysisError {
@@ -18,35 +14,5 @@ pub enum AnalysisError {
     Execution(#[from] ExecutionError),
 
     #[error("Comparison failed for case '{1}': {0}")]
-    ComparisonFailed(ComparisonResult, String),
-}
-
-impl AnalysisContext {
-    pub fn new(
-        testname: String,
-        gccrs_binary: &Path,
-        rustc_binary: &Path,
-        timeout: Duration,
-    ) -> Self {
-        Self {
-            testname,
-            gccrs: ExecutionContext::new(gccrs_binary, timeout),
-            rustc: ExecutionContext::new(rustc_binary, timeout),
-        }
-    }
-
-    pub fn analyze(&self) -> Result<(), AnalysisError> {
-        let gccrs_exec_result = self.gccrs.run_binary()?;
-        let rustc_exec_result = self.rustc.run_binary()?;
-        let compare = Comparison::new(gccrs_exec_result, rustc_exec_result);
-        let result = compare.compare();
-        if result.is_identical() {
-            Ok(())
-        } else {
-            Err(AnalysisError::ComparisonFailed(
-                result,
-                self.testname.clone(),
-            ))
-        }
-    }
+    ComparisonFailed(AnalysisResult, String),
 }
